fix(ses,athena): typed-client coverage slice 18 — five services fully covered, five wire bugs
gopherstack-n3zi slice 18. ses 44 → 0 uncovered (71/71), emr 44 → 0
(65/65), athena 51 → 0 (70/70), workspaces 51 → 0 (91/91), appstream
48 → 0 (89/89); repo-wide 8589/10856 (79%).

ses SendRawEmail never base64-decoded RawMessage.Data, which the query
serializer always base64-encodes (awsAwsquery_serializeDocumentRawMessage)
— every real raw message failed to parse; the decode is tolerant so
hand-built MIME tests still pass. GetCustomVerificationEmailTemplate
wrapped its members in a nonexistent element; the real output is flat.
athena CreatePresignedNotebookUrl's AuthTokenExpirationTime is a plain
int64, not a timestamp, and the fractional epoch broke ParseInt;
Executor Start/TerminationDateTime were float64 where the real fields
are int64; StartSession has no top-level NotebookId — it travels in
EngineConfiguration.AdditionalConfigs — so Session.NotebookID was always
empty and ListNotebookSessions never matched. emr, workspaces and
appstream decoded cleanly. emr ListTagsForResource (no wire op),
workspaces TerminateWorkspacesPoolSession (no producing path) and
appstream DescribeAppLicenseUsage nil-list are recorded. No persisted
change.

+++ services/athena/models.go
@@ -533,14 +533,18 @@ type CapacityAssignmentConfiguration struct {
 	CapacityAssignments     []CapacityAssignment `json:"CapacityAssignments,omitempty"`
 }
 
-// Executor describes a Spark executor.
+// Executor describes a Spark executor. StartDateTime/TerminationDateTime are
+// plain epoch-seconds *int64 on the real wire (athena@v1.60.4 types.go's
+// ExecutorsSummary), NOT a smithy timestamp -- their deserializer calls
+// strconv.ParseInt, which errors on a fractional value, unlike the
+// time.Time-typed epoch fields elsewhere in this package.
 type Executor struct {
-	ExecutorID          string  `json:"ExecutorId"`
-	ExecutorType        string  `json:"ExecutorType"`
-	ExecutorState       string  `json:"ExecutorState"`
-	StartDateTime       float64 `json:"StartDateTime,omitempty"`
-	TerminationDateTime float64 `json:"TerminationDateTime,omitempty"`
-	ExecutorSize        int64   `json:"ExecutorSize,omitempty"`
+	ExecutorID          string `json:"ExecutorId"`
+	ExecutorType        string `json:"ExecutorType"`
+	ExecutorState       string `json:"ExecutorState"`
+	StartDateTime       int64  `json:"StartDateTime,omitempty"`
+	TerminationDateTime int64  `json:"TerminationDateTime,omitempty"`
+	ExecutorSize        int64  `json:"ExecutorSize,omitempty"`
 }
 
 // EngineVersionDescriptor describes an available engine version.
